% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2 functions.R
\docType{data}
\name{A3}
\alias{A3}
\title{Question 3 and 4}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
A3
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Question 3 and 4
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
stand(A3)
}
\keyword{datasets}
