% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2 functions.R
\name{func8}
\alias{func8}
\title{Question 1}
\usage{
func8(a, x)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Question 1
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
func8(a,x)
}
