% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2 functions.R
\name{\%alice\%}
\alias{\%alice\%}
\title{Question 2}
\usage{
a \%alice\% x
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Question 2
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
different version
}
