% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2 functions.R
\name{myapply}
\alias{myapply}
\title{Question 5}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{x}{vector}
}
\value{
list
}
\description{
Question 5
}
\examples{
fred <-matrix(1:6, ncol = 2)
myapply(fred,2,"quantile",prob = 0.75)
myapply(fred,1,"quantile",prob = (1:3)/4)
}
