% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{plot1}
\alias{plot1}
\title{Wrapper for ggplot2 for data d}
\usage{
plot1(x, p)
}
\arguments{
\item{x}{data.frame}
}
\value{
ggplot2
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
data(d)
plot1(d$x,d$p)
}
